drop buf2rgb and move functions to classes
