+++ drmdevice.cpp
@@ -56,7 +56,6 @@ extern "C" {
 }
 
 #include "misc.h"
-#include "buf2rgb.h"
 
 #include "videorender.h"
 #include "audio.h"
