+++ softhddevice.cpp
@@ -53,8 +53,6 @@ extern "C" {
 #include <libavutil/timestamp.h>
 
 }
-#include "buf2rgb.h"
-
 #include "iatomic.h"
 #include "videostream.h"
 #include "audio.h"
