+++ videostream.cpp
@@ -44,8 +44,6 @@ extern "C" {
 #include "logger.h"
 #include "h264parser.h"
 
-#include "buf2rgb.h"
-
 #include "iatomic.h"
 #include "videostream.h"
 #include "audio.h"
